% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanet.R
\name{sanet}
\alias{sanet}
\title{This function takes in a dataframe with 'node' and 'activation' columns, and
calls the spread.parallel() function for a specific number of times to simulate the spread of
activation in a network.}
\usage{
sanet(start_run, decay = 0.2, retention = 0.8, suppress = 0.1, network,
  time = 10)
}
\arguments{
\item{start_run}{A non-empty dataframe with 'node' and 'activation' columns. Must be specified.}

\item{decay}{Proportion of activation that is lost at each time step. Ranges from 0 to 1. Default is 0.2.}

\item{retention}{Proportion of activation that remains in the node from which activation is being spread from. Ranges from 0 to 1. Default is 0.8.}

\item{suppress}{Suppress nodes with a final activation of < x units at each time step to an activation value of 0. Suggested value of x is 0.1 percent of total initial activation at t = 0 in order to speed up simulations substantially. Default is 0.1.}

\item{network}{Network where the spreading occurs. Must be specified. Must be an igraph object with a "name" property.}

\item{time}{Number of time steps to run the spread() function for. Default is 10.}
}
\value{
A compiled dataframe with 'node', 'activation' and 'time' columns showing the spread of activation in the network over time.
}
\description{
This function takes in a dataframe with 'node' and 'activation' columns, and
calls the spread.parallel() function for a specific number of times to simulate the spread of
activation in a network.
}
\examples{
See Vignette for examples.
}
