% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread.parallel.R
\name{spread.parallel}
\alias{spread.parallel}
\title{This function takes in a dataframe with 'node' and 'activation' columns, and
simulates the spread of activation of nodes listed in the 'node' column to its
neighbors in the specified network.}
\usage{
spread.parallel(start_run, decay, retention, suppress, network)
}
\arguments{
\item{start_run}{A non-empty dataframe with 'node' and 'activation' columns.}

\item{decay}{Proportion of activation that is lost at each time step. Ranges from 0 to 1.}

\item{retention}{Proportion of activation that remains in the node from which activation is being spread from. Ranges from 0 to 1.}

\item{suppress}{Suppress nodes with a final activation of < x units at each time step to an activation value of 0. Suggested value of x is 0.1 percent of total initial activation at t = 0 in order to speed up simulations substantially.}

\item{network}{Network where the spreading occurs.}
}
\value{
A updated dataframe with 'node' and 'activation' columns with new activation values.
}
\description{
This function takes in a dataframe with 'node' and 'activation' columns, and
simulates the spread of activation of nodes listed in the 'node' column to its
neighbors in the specified network.
}
\examples{
See Vignette for examples.
}
